# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready blog backend API for Alfafaa Community - a Muslim community blogging platform.",
        "title": "Alfafaa Blog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alfafaa Support",
            "url": "https://alfafaa.com/support",
            "email": "support@alfafaa.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/articles": {
            "get": {
                "description": "Get a paginated list of published articles (editors can see all)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "List articles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category slug",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by tag slug",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by author ID",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in title and content",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new article (requires author role or higher)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Create article",
                "parameters": [
                    {
                        "description": "Article data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Article created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires author role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/articles/recent": {
            "get": {
                "description": "Get the most recently published articles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get recent articles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of articles to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent articles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/articles/trending": {
            "get": {
                "description": "Get articles sorted by view count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get trending articles",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of articles to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trending articles retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an article (authors can only update their own, editors can update any)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Update article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Article update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not the author",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an article (authors can only delete their own, editors can delete any)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Delete article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - not the author",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/articles/{id}/publish": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Publish an article (requires editor role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Publish article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article published successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires editor role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
        "/articles/{id}/unpublish": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unpublish an article (requires editor role)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Unpublish article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article unpublished successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.ArticleDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires editor role",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_alfafaa_alfafaa-blog_internal_dto.ArticleListResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.PublicUserResponse"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.CategoryResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "featured_image_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "reading_time_minutes": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_dto.TagResponse"
                    }
                },
                "title": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_alfafaa_alfafaa-blog_internal_utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_alfafaa_alfafaa-blog_internal_utils.ResponseWithMeta": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_alfafaa_alfafaa-blog_internal_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}